+++ preload.js
@@ -7,6 +7,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   getVideoInfo: (videoId) => ipcRenderer.invoke('get-video-info', videoId),
   getChannelVideos: (channelId) => ipcRenderer.invoke('get-channel-videos', channelId),
   getPlaylistVideos: (playlistId) => ipcRenderer.invoke('get-playlist-videos', playlistId),
+  getTrendingVideos: () => ipcRenderer.invoke('get-trending-videos'),
   minimizeWindow: () => ipcRenderer.send('minimize-window'),
   closeWindow: () => ipcRenderer.send('close-window'),
   toggleAlwaysOnTop: () => ipcRenderer.invoke('toggle-always-on-top'),